add diffusers to coverage test
